requirements.txt
Django==2.2.24
djangorestframework==3.12.4
drf-writable-nested==0.6.3
drf-yasg==1.21.7
pytest==7.4.4
pytest-django==4.7.0
gunicorn==20.1.0
whitenoise==5.3.0
